+++ src/main.tsx
@@ -6,6 +6,7 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import './themes/tokens.css';
+import './themes/eraThemes.css';
 import './styles/index.css';
 import { App } from './App';